# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VFADigital Engine - Core API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  model: "model",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatRequest {
  name: string;
  phone?: string;
  message: string;
  history?: ChatMessage[];
}

export interface ChatResponse {
  reply: string;
  summary?: string;
}

export interface AdminVerifyRequest {
  token: string;
}

export type AdminDataSettings = { [key: string]: unknown };

export interface AdminData {
  settings: AdminDataSettings;
}

export interface AdminVerifyResponse {
  success: boolean;
  data: AdminData;
}

export interface ContactSettings {
  whatsapp: string;
  facebook: string;
  instagram: string;
  github: string;
}

export interface PortfolioItem {
  id: string;
  name: string;
  description: string;
  tags: string[];
}

export interface StatsSettings {
  projects: number;
  clients: number;
  experience: number;
}

export interface PublicSettings {
  contact: ContactSettings;
  portfolio: PortfolioItem[];
  stats: StatsSettings;
}

export interface ErrorResponse {
  error: string;
}

export type GetLocale200 = { [key: string]: string };