# lib/api-zod/src/generated/types/adminData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VFADigital Engine - Core API
 * OpenAPI spec version: 1.0.0
 */
import type { AdminDataSettings } from "./adminDataSettings";

export interface AdminData {
  settings: AdminDataSettings;
}