# lib/api-zod/src/generated/types/adminVerifyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VFADigital Engine - Core API
 * OpenAPI spec version: 1.0.0
 */
import type { AdminData } from "./adminData";

export interface AdminVerifyResponse {
  success: boolean;
  data: AdminData;
}