# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VFADigital Engine - Core API
 * OpenAPI spec version: 1.0.0
 */
import type { ChatMessage } from "./chatMessage";

export interface ChatRequest {
  name: string;
  phone?: string;
  message: string;
  history?: ChatMessage[];
}