# lib/api-zod/src/generated/types/portfolioItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VFADigital Engine - Core API
 * OpenAPI spec version: 1.0.0
 */

export interface PortfolioItem {
  id: string;
  name: string;
  description: string;
  tags: string[];
}