# lib/api-zod/src/generated/types/publicSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VFADigital Engine - Core API
 * OpenAPI spec version: 1.0.0
 */
import type { ContactSettings } from "./contactSettings";
import type { PortfolioItem } from "./portfolioItem";
import type { StatsSettings } from "./statsSettings";

export interface PublicSettings {
  contact: ContactSettings;
  portfolio: PortfolioItem[];
  stats: StatsSettings;
}